feat: move utility extensions and small helpers to commons
- Move togglePresenceInSet (SetExt) to commons/util/SetUtils.kt
- Move isLiveStreaming to commons/util/IsLiveStreaming.kt
- Convert IterableExt.kt to backward-compat wrapper (already in commons)
- Convert ByteFormatter.kt to backward-compat wrapper (already in commons)
- Convert CountFormatter.kt to backward-compat wrapper (already in commons)

All original files now re-export from commons for backwards compatibility.
