+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/util/IsLiveStreaming.kt
@@ -18,13 +18,6 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.service
+package com.vitorpamplona.amethyst.commons.util
 
-// Re-export from commons for backwards compatibility
-fun countToHumanReadableBytes(counter: Int) =
-    com.vitorpamplona.amethyst.commons.util
-        .countToHumanReadableBytes(counter)
-
-fun countToHumanReadableBytes(counter: Long) =
-    com.vitorpamplona.amethyst.commons.util
-        .countToHumanReadableBytes(counter)
+fun isLiveStreaming(url: String) = url.contains(".m3u8", true)
